spring-mvc-rest-app/src/main/java/tech/talci/services/CategoryServiceImpl.java
package tech.talci.services;

import lombok.Setter;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;
import tech.talci.api.v1.mapper.CategoryMapper;
import tech.talci.api.v1.model.CategoryDTO;
import tech.talci.repositories.CategoryRepository;

import java.util.List;
import java.util.stream.Collectors;

@Service
@Slf4j
public class CategoryServiceImpl implements CategoryService {

    private CategoryRepository categoryRepository;
    private final CategoryMapper categoryMapper;

    public CategoryServiceImpl(CategoryRepository categoryRepository, CategoryMapper categoryMapper) {
        this.categoryRepository = categoryRepository;
        this.categoryMapper = categoryMapper;
    }

    @Override
    public List<CategoryDTO> getAllCategories() {
        log.debug("Getting list of all categories");

        return categoryRepository
                .findAll()
                .stream()
                .map(categoryMapper::categoryToCategoryDTO)
                .collect(Collectors.toList());
    }

    @Override
    public CategoryDTO getCategoryByName(String name) {

        log.debug("Getting category by name: " + name);

        return categoryMapper.categoryToCategoryDTO(categoryRepository.findByName(name));
    }
}
